pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "colabfit-mcp"
version = "0.1.0"
description = "MCP tools for ColabFit dataset discovery, model training and deployment"
license = {file = "LICENSE"}
readme = "README.md"
authors = [{name = "ColabFit"}]
requires-python = ">= 3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "requests",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Utilities",
    "License :: OSI Approved :: BSD License",
]
keywords = [
    "machine learning",
    "interatomic potentials",
    "MCP",
    "model context protocol",
    "materials science",
]

[project.optional-dependencies]
full = [
    "klay @ git+https://github.com/openkim/klay.git",
    "kliff @ git+https://github.com/openkim/kliff.git",
    "pytorch-lightning>=2.0",
    "torch_geometric",
    "torch_scatter",
    "torch_ema",
    "kimpy",
    "ase",
    "datasets>=2.0",
    "huggingface_hub",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
colabfit-mcp = "colabfit_mcp:main"

[project.urls]
Homepage = "https://github.com/colabfit/colabfit-mcp"
